+++ src/pybel/io/__init__.py
@@ -31,6 +31,7 @@ from .nodelink import (
     from_nodelink, from_nodelink_file, from_nodelink_gz, from_nodelink_jsons, to_nodelink,
     to_nodelink_file, to_nodelink_gz, to_nodelink_jsons,
 )
+from .pynpa import to_npa_dfs, to_npa_directory
 from .tsv import to_edgelist, to_tsv
 from .umbrella_nodelink import to_umbrella_nodelink, to_umbrella_nodelink_file, to_umbrella_nodelink_gz
 from .web import from_web, to_web